pyproject.toml
[project]
name = "aituber-build-tool"
version = "0.1.0"
description = "AITuber framework for virtual characters"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.2.0",
    "openai>=1.0.0",
    "chromadb>=0.4.13",
    "pyyaml>=6.0.2",
    "typer>=0.15.2",
]

[project.scripts]
aituber = "aituber.cli.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aituber"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.7",
]
